Recommendation page, reformatted links to be using next_book/*, still working with sending params to this servlet to add stuff to the database

+++ book_web_app/src/main/java/com/nextBook/database/entity/Book.java
@@ -82,7 +82,7 @@ public class Book extends BaseEntity {
      * Constructor for non-null fields. Testing purposes mainly.
      */
     public Book(String bookTitle, String anAuthorName,
-                int aBookRating, List<String> bookGenres, String anIsbn) {
+                double aBookRating, List<String> bookGenres, String anIsbn) {
         this();
         //this.id = anId;
         this.title = bookTitle;
